Split functions for character profiles and searching into their own files
